requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
celery>=5.3.4
redis>=5.0.1
django-redis>=5.4.0
djangorestframework>=3.14.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0
Pillow>=10.1.0
pandas>=2.1.3
openpyxl>=3.1.2
django-cors-headers>=4.3.1
django-filter>=23.5
requests>=2.31.0
celery[redis]>=5.3.4
flower>=2.0.1
django-extensions>=3.2.3
django-celery-results>=2.5.1
django-celery-beat>=2.5.0
python-dateutil>=2.8.2
pytz>=2023.3
